% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amelioration-annotation.R
\name{amelioration.annotation}
\alias{amelioration.annotation}
\title{Title}
\usage{
amelioration.annotation(species, collicin, annotationDir)
}
\arguments{
\item{annotationDir}{}
}
\value{

}
\description{
Title
}
